Moved models and common utilities to a separate class library.
