+++ ConfigUtil.sln
@@ -19,6 +19,10 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "ConfigUtil", "src\ConfigUti
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Launcher", "src\Launcher\Launcher.csproj", "{514F2642-3881-47DA-8141-65784EDAE8DA}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "OSVR.Config", "src\OSVR.Config\OSVR.Config.xproj", "{DCA309D3-AFAF-4C03-8AA0-2F29A5F09842}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "OSVR.Config.Controllers", "src\OSVR.Config.Controllers\OSVR.Config.Controllers.xproj", "{9AEE08DE-4CE5-453B-8AFC-0F091AC808DB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +37,14 @@ Global
 		{514F2642-3881-47DA-8141-65784EDAE8DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{514F2642-3881-47DA-8141-65784EDAE8DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{514F2642-3881-47DA-8141-65784EDAE8DA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DCA309D3-AFAF-4C03-8AA0-2F29A5F09842}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DCA309D3-AFAF-4C03-8AA0-2F29A5F09842}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DCA309D3-AFAF-4C03-8AA0-2F29A5F09842}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DCA309D3-AFAF-4C03-8AA0-2F29A5F09842}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9AEE08DE-4CE5-453B-8AFC-0F091AC808DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9AEE08DE-4CE5-453B-8AFC-0F091AC808DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9AEE08DE-4CE5-453B-8AFC-0F091AC808DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9AEE08DE-4CE5-453B-8AFC-0F091AC808DB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -40,5 +52,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{B2D61055-0BB3-46DE-80F2-17EF716B8746} = {9C3EC537-B59E-4121-A0CA-F16CDC241CB4}
 		{514F2642-3881-47DA-8141-65784EDAE8DA} = {9C3EC537-B59E-4121-A0CA-F16CDC241CB4}
+		{DCA309D3-AFAF-4C03-8AA0-2F29A5F09842} = {9C3EC537-B59E-4121-A0CA-F16CDC241CB4}
+		{9AEE08DE-4CE5-453B-8AFC-0F091AC808DB} = {9C3EC537-B59E-4121-A0CA-F16CDC241CB4}
 	EndGlobalSection
 EndGlobal
