+++ src/ConfigUtil/Controllers/CurrentConfigController.cs
@@ -21,10 +21,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using ConfigUtil.Models;
+using OSVR.Config.Models;
 using Microsoft.Extensions.Configuration;
 using System.IO;
-using ConfigUtil.Common;
+using OSVR.Config.Common;
 using Newtonsoft.Json;
 
 namespace ConfigUtil.Controllers
