+++ src/ConfigUtil/Controllers/EnableDirectModeController.cs
@@ -18,8 +18,8 @@
 /// 
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
-using ConfigUtil.Common;
-using ConfigUtil.Models;
+using OSVR.Config.Common;
+using OSVR.Config.Models;
 
 namespace ConfigUtil.Controllers
 {
