+++ src/ConfigUtil/project.json
@@ -1,4 +1,4 @@
-﻿{
+{
   "dependencies": {
     "Microsoft.NETCore.App": "1.0.0",
     "Microsoft.AspNetCore.Mvc": "1.0.0",
@@ -10,7 +10,8 @@
     "Microsoft.Extensions.Logging.Console": "1.0.0",
     "Microsoft.Extensions.Logging.Debug": "1.0.0",
     "Microsoft.Extensions.Options.ConfigurationExtensions": "1.0.0",
-    "Microsoft.AspNetCore.StaticFiles":  "1.0.0"
+    "Microsoft.AspNetCore.StaticFiles": "1.0.0",
+    "OSVR.Config": "1.0.0-*"
   },
 
   "frameworks": {
